Move VideoPlayer into its own top level class
